Congratulations, the terrain Works™️ after only too long.

+++ src/consts.rs
@@ -48,13 +48,13 @@ pub const TIMEOUT_DURATION: std::time::Duration = std::time::Duration::from_secs
 pub const AUTOSAVE_FREQUENCY: std::time::Duration = std::time::Duration::from_secs(60 * 5);
 
 /// Size of the player hitbox in pixels
-pub const PLAYER_HITBOX: (f64, f64) = (64.0, 64.0);
+pub const PLAYER_HITBOX: (f32, f32) = (64.0, 64.0);
 /// The distance before an item on the ground is drawn to players
 pub const ITEM_MAGNET_DISTANCE: f32 = 256.0;
 /// The distance before an item on the ground is picked up by players
 pub const ITEM_PICKUP_DISTANCE: f32 = 16.0;
 /// The distance a player can be from a tree and still successfully hit it when chopping
-pub const TREE_CHOP_DISTANCE: f64 = (PLAYER_HITBOX.0 / 2.0) + 64.0;
+pub const TREE_CHOP_DISTANCE: f32 = (PLAYER_HITBOX.0 / 2.0) + 64.0;
 /// The offset between the cursor's render location and actual location
 pub const CURSOR_OFFSET: [f32; 2] = [-25.0, 31.0];
 /// Width of a chunk in tiles
